requirements.txt
streamlit==1.38.0
google-cloud-aiplatform==1.65.0
google-api-python-client==2.149.0
google-auth==2.35.0
chromadb==0.5.5
langchain==0.2.16
langchain-text-splitters==0.2.2
vertexai==1.65.0
tiktoken==0.7.0
protobuf==4.25.3
requests==2.32.3
pysqlite3-binary==0.5.3

